# 1869: Course program api performance improvements