--- a/courses/serializers.py
+++ b/courses/serializers.py
@@ -152,8 +152,6 @@ class CourseRunSerializer(BaseCourseRunSerializer):
 class DepartmentSerializer(serializers.ModelSerializer):
     """Department model serializer"""
 
-    name = CharField(max_length=128)
-
     class Meta:
         model = models.Department
         fields = ["name"]