… courses/migrations/0043_alter_course_live.py
… courses/migrations/0043_course_program_live_index.py
… courses/migrations/0044_alter_program_live.py
… courses/serializers_test.py
… courses/views_test.py
… frontend/public/src/components/EnrolledItemCard.js
… frontend/public/src/containers/ProductDetailEnrollApp.js
… frontend/public/src/containers/ProductDetailEnrollApp_test.js
… frontend/public/src/containers/UpsellCardApp.js
… frontend/public/src/containers/pages/CatalogPage.js
… frontend/public/src/containers/pages/CatalogPage_test.js
… frontend/public/src/factories/course.js
… frontend/public/src/lib/courseApi.js
… frontend/public/src/lib/courseApi_test.js
… frontend/public/src/lib/queries/courseRuns.js
… frontend/public/src/lib/queries/programs.js